test: share the cookie-inventory guard, and apply it to the GVL twin
gvl-vendor-auto-detect is the twin of cookie-policy-service-auto-detect: same
resetCookies() deleting only its own fixture slugs, same assertions about the
WHOLE inventory (vendor_ids is [], already_selected is exactly [755],
newly_suggested is exactly [804]). It carried the identical latent defect and
surfaced five reds in the last full run — the same shape I had just fixed next
door, in the spec I had not looked at.

The guard now lives in tests/e2e/utils/cookie-inventory.ts instead of inside one
spec. That placement is the point: the original defect was two hand-written
copies of the same snapshot/wipe/restore inside a single file, which could not
stay in step and left eight sibling tests unguarded. A second spec needing the
same thing is exactly the case a local helper fails.

Four tests wrapped in the GVL spec (3, 5, 7, 8) and the cookie-policy spec moved
onto the shared helper. Test 8 was missed on the first pass because its expect()
carries a message argument, so the pattern that found the others did not match
it — worth stating, since the same blind spot will recur.

Verified the way the last one had to be: both specs run together with four
foreign discovered rows planted (.googletagmanager.com, .linkedin.com,
.tiktok.com, .youtube.com) and confirmed still present afterwards — 36/36.

The other two reds from that run, blocking-compliance:325 and setup-wizard:272,
pass in isolation (21/21). That run took 2.1h against the usual 55m with 138
flaky, on a box whose load average was 7.66 with Spotlight reindexing at 101%
CPU; its numbers are not a fair measurement of anything.
